requirements.txt
asgiref==3.6.0
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
channels==3.0.0
charset-normalizer==3.0.1
constantly==15.1.0
cryptography==39.0.1
daphne==3.0.2
Django==4.1.6
django-grip==3.2.0
gripcontrol==4.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
MarkupSafe==2.1.2
pubcontrol==3.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
python-dotenv==0.21.1
python-owasp-zap-v2.4==0.0.21
requests==2.28.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
zope.interface==5.5.2
